Feature/changes v1 (#1)
* changes in architecture
