--- cmd/builder/consumer.go
@@ -1,26 +1,0 @@
-package builder
-
-import (
-	"fmt"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"go.uber.org/zap"
-	"service-worker-sqs-postgres/clients/awssqs"
-	"service-worker-sqs-postgres/consumer"
-	"service-worker-sqs-postgres/database"
-	"service-worker-sqs-postgres/domain"
-)
-
-// NewSQS define all services to instantiate SQS.
-func NewSQS(logger *zap.SugaredLogger, config *Configuration, sessionaws *session.Session, db *database.ClientDB) (domain.Source, error) {
-	sqs, err := awssqs.NewSQSClient(sessionaws, config.SQSUrl, config.SQSMaxMessages, config.SQSVisibilityTimeout)
-	if err != nil {
-		return nil, fmt.Errorf("error awssqs.NewSQSClient: %w", err)
-	}
-
-	source, err := consumer.New(sqs, logger, config.SQSMaxMessages, db)
-	if err != nil {
-		return nil, fmt.Errorf("error consumer.New: %w", err)
-	}
-
-	return source, nil
-}
