+++ core/domain/event.go
