--- http/controllers/events.go
@@ -1,34 +1,0 @@
-package controllers
-
-import (
-	"github.com/labstack/echo"
-	"net/http"
-	"service-worker-sqs-postgres/domain/entity"
-	"service-worker-sqs-postgres/http/services"
-	env "service-worker-sqs-postgres/utils"
-)
-
-// EventsController encapsulates all the data necessary for the implementation of the EventsService.
-type EventsController struct {
-	eventsService services.EventsService
-}
-
-// NewEventsController instantiate a new event controller.
-func NewEventsController(es services.EventsService) *EventsController {
-	return &EventsController{
-		eventsService: es,
-	}
-}
-
-// GetID return a event by ID [eventsService.GetID].
-func (ec *EventsController) GetID(c echo.Context) error {
-	ID, err := env.GetParam(c, "id")
-	if err != nil {
-		return entity.NewError(http.StatusBadRequest, err)
-	}
-	events, err := ec.eventsService.GetID(ID)
-	if err != nil {
-		return entity.HandleServiceError(err)
-	}
-	return c.JSON(http.StatusOK, events.ToDomainEvents())
-}
